Reorganize version, utils, and experimental modules
