+++ docs/api/experimental.rst
@@ -1,0 +1,5 @@
+phasorpy.experimental
+---------------------
+
+.. automodule:: phasorpy.experimental
+    :members:
