+++ tests/test_cli.py
@@ -2,8 +2,9 @@
 
 from click.testing import CliRunner
 
-from phasorpy import __version__, versions
+from phasorpy import __version__
 from phasorpy.cli import main
+from phasorpy.utils import versions
 
 
 def test_version():
