+++ tests/test_phasorpy.py
@@ -1,17 +1,11 @@
 """Tests for the phasorpy module."""
 
-from phasorpy import __version__, versions
 
+def test_version():
+    """Test phasorpy.__version__ attribute."""
+    from phasorpy import __version__
 
-def test_versions():
-    """Test phasorpy.versions function."""
-    ver = versions()
-    assert 'Python-' in ver
-    assert f'phasorpy-{__version__}\nnumpy-' in ver
-    assert '(' not in ver
-
-    ver = versions(sep=', ', dash=' ', verbose=True)
-    assert f', phasorpy {__version__}  (' in ver
+    assert '.' in __version__
 
 
 # mypy: allow-untyped-defs, allow-untyped-calls
